+++ app-backend/batch/src/main/scala/com/azavea/rf/batch/ingest/spark/Ingest.scala
@@ -1,8 +1,9 @@
-package com.azavea.rf.batch.ingest
+package com.azavea.rf.batch.ingest.spark
 
 import io.circe.parser._
 
 import com.azavea.rf.batch._
+import com.azavea.rf.batch.ingest._
 import com.azavea.rf.batch.ingest.model._
 import com.azavea.rf.batch.util._
 
@@ -144,7 +145,8 @@ object Ingest extends SparkJob with LazyLogging {
         FileRangeReader(new File(uri))
       case "s3" | "s3a" | "s3n" =>
         val (bucket, prefix) = S3.parse(uri)
-        S3RangeReader(bucket, prefix, S3Client.DEFAULT)
+        val decodedPrefix = java.net.URLDecoder.decode(prefix, "UTF-8")
+        S3RangeReader(bucket, decodedPrefix, S3Client.DEFAULT)
       case "http" | "https"
           if uri.getAuthority == "s3.amazonaws.com"
           && uri.getQuery.contains("AWSAccessKeyId") =>