+++ app-backend/batch/src/main/scala/com/azavea/rf/batch/util/S3.scala
@@ -1,30 +1,32 @@
 package com.azavea.rf.batch.util
 
 import geotrellis.spark.io.s3.S3InputFormat
-
 import com.amazonaws.auth.DefaultAWSCredentialsProviderChain
 import com.amazonaws.services.s3.AmazonS3ClientBuilder
-import com.amazonaws.services.s3.model.{ListObjectsRequest, ObjectListing, S3Object}
+import com.amazonaws.services.s3.model.{ListObjectsRequest, ObjectListing, PutObjectResult, S3Object}
 import org.apache.hadoop.conf.Configuration
-
 import java.net.URI
 
 import scala.collection.mutable
 import scala.collection.JavaConverters._
 
-object S3 {
-  def getClient(region: Option[String] = None) = {
-    val builder = AmazonS3ClientBuilder.standard()
-      .withCredentials(new DefaultAWSCredentialsProviderChain())
+case class S3(region: Option[String] = None) extends Serializable {
+  lazy val client = {
+    val builder =
+      AmazonS3ClientBuilder
+        .standard()
+        .withCredentials(new DefaultAWSCredentialsProviderChain())
+
 
     region.fold(builder)(builder.withRegion).build()
   }
 
-  lazy val client = getClient()
-
   /** Get S3Object */
-  def getObject(s3bucket: String, s3prefix: String, region: Option[String] = None): S3Object =
-    getClient(region).getObject(s3bucket, s3prefix)
+  def getObject(s3bucket: String, s3prefix: String): S3Object =
+    client.getObject(s3bucket, s3prefix)
+
+  def putObject(s3bucket: String, s3Key: String, content: String): PutObjectResult =
+    client.putObject(s3bucket, s3Key, content)
 
   /** List the keys to files found within a given bucket */
   def listKeys(url: String, ext: String, recursive: Boolean): Array[URI] = {
@@ -59,9 +61,11 @@ object S3 {
       listObjectsRequest.setMarker(listing.getNextMarker)
     } while (listing.isTruncated)
 
-    result.toSeq
+    result
   }
+}
 
+object S3 {
   /** Parse an S3 URI unto its bucket and prefix portions */
   def parse(uri: URI): (String, String) = {
     val S3InputFormat.S3UrlRx(_, _, bucket, prefix) = uri.toString