+++ app-tasks/rf/src/rf/ingest/geotiff_ingest.py
@@ -7,6 +7,9 @@ from .models import Ingest, Source, Layer
 
 layer_s3_bucket = os.getenv('TILE_SERVER_BUCKET')
 
+# 1 cm resolution is as high as we go, otherwise zoom levels get ridiculous
+MIN_RESOLUTION_METERS = .1
+
 
 def get_safe_uri(uri):
     """Parse URI and ensure the path is escaped properly
@@ -40,8 +43,15 @@ def get_source_definition(image, extent, crs=None):
     """
 
     uri = get_safe_uri(image.sourceUri)
+
     band_maps = [{'source': band.number, 'target': band.number} for band in image.bands]
-    cell_size = {'width': image.resolutionMeters, 'height': image.resolutionMeters}
+    if image.resolutionMeters < MIN_RESOLUTION_METERS:
+        width = MIN_RESOLUTION_METERS
+        height = MIN_RESOLUTION_METERS
+    else:
+        width = image.resolutionMeters
+        height = image.resolutionMeters
+    cell_size = {'width': height, 'height': width}
     extent_crs = 'epsg:4326'
     return Source(uri, extent, band_maps, cell_size, extent_crs, crs)
 
@@ -58,7 +68,14 @@ def get_ingest_layer(scene):
     extent = scene.get_extent()
     sources = [get_source_definition(image, extent) for image in scene.images]
     highest_resolution_meters = min([image.resolutionMeters for image in scene.images])
-    cell_size = {'width': highest_resolution_meters, 'height': highest_resolution_meters}
+    if highest_resolution_meters < MIN_RESOLUTION_METERS:
+        width = MIN_RESOLUTION_METERS
+        height = MIN_RESOLUTION_METERS
+    else:
+        width = highest_resolution_meters
+        height = highest_resolution_meters
+
+    cell_size = {'width': width, 'height': height}
     output_uri = 's3://{}/layers'.format(layer_s3_bucket)
     return Layer(scene.id, output_uri, sources, cell_size)